Add tracking for active tab (browser tab)
With this change messages like "match_assigned" will only be send
to the currently active tab. When the user closes a tab another
will be set to be the active one.
This change fixed the problem, that a user gets multiple messages
that a match is ready and because of that every tab could open a
game (When someone wants to abuse this, it would still be possible
but much harder for a normal user)